21.2

+++ src/pipes.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   pipes.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fmarin-p <fmarin-p@student.42.fr>          +#+  +:+       +#+        */
+/*   By: clcarrer <clcarrer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/17 13:44:44 by fmarin-p          #+#    #+#             */
-/*   Updated: 2023/02/19 10:56:17 by fmarin-p         ###   ########.fr       */
+/*   Updated: 2023/02/21 16:32:31 by clcarrer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@ void	red_pipe_child(t_cmdtable *rl, int i)
 {
 	if (i == 0 && rl->infile)
 		dup2(rl->infile, 0);
-	if ((i == (int)rl->n_cmd - 1) && rl->outfile)
+	if ((i == rl->n_cmd - 1) && rl->outfile)
 		dup2(rl->outfile, 1);
-	if (i != (int)rl->n_cmd - 1)
+	if (i != rl->n_cmd - 1)
 	{
 		close(rl->pipe[0]);
 		dup2(rl->pipe[1], 1);
